tfmini split out header and main
