+++ src/drivers/distance_sensor/tfmini/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2017 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2017-2019 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -33,10 +33,10 @@
 px4_add_module(
 	MODULE drivers__tfmini
 	MAIN tfmini
-	COMPILE_FLAGS
-		-Wno-cast-align # TODO: fix and enable
 	SRCS
-		tfmini.cpp
+		TFMINI.cpp
+		TFMINI.hpp
+		tfmini_main.cpp
 		tfmini_parser.cpp
 	MODULE_CONFIG
 		module.yaml
