+++ src/drivers/distance_sensor/tfmini/TFMINI.cpp
@@ -31,72 +31,7 @@
  *
  ****************************************************************************/
 
-/**
- * @file tfmini.cpp
- * @author Lorenz Meier <lm@inf.ethz.ch>
- * @author Greg Hulands
- * @author Ayush Gaud <ayush.gaud@gmail.com>
- * @author Christoph Tobler <christoph@px4.io>
- * @author Mohammed Kabir <mhkabir@mit.edu>
- *
- * Driver for the Benewake TFmini laser rangefinder series
- */
-
-#include <termios.h>
-
-#include <drivers/drv_hrt.h>
-#include <perf/perf_counter.h>
-#include <px4_config.h>
-#include <px4_getopt.h>
-#include <px4_module.h>
-#include <px4_platform_common/px4_work_queue/ScheduledWorkItem.hpp>
-#include <drivers/rangefinder/PX4Rangefinder.hpp>
-#include <uORB/topics/distance_sensor.h>
-
-#include "tfmini_parser.h"
-
-#define TFMINI_DEFAULT_PORT	"/dev/ttyS3"
-
-using namespace time_literals;
-
-class TFMINI : public px4::ScheduledWorkItem
-{
-public:
-	TFMINI(const char *port, uint8_t rotation = distance_sensor_s::ROTATION_DOWNWARD_FACING);
-	virtual ~TFMINI();
-
-	int init();
-
-	void print_info();
-
-private:
-
-	int collect();
-
-	void Run() override;
-
-	void start();
-	void stop();
-
-	PX4Rangefinder	_px4_rangefinder;
-
-	TFMINI_PARSE_STATE _parse_state {TFMINI_PARSE_STATE::STATE0_UNSYNC};
-
-	char _linebuf[10] {};
-	char _port[20] {};
-
-	static constexpr int kCONVERSIONINTERVAL{9_ms};
-
-	int _fd{-1};
-
-	unsigned int _linebuf_index{0};
-
-	hrt_abstime _last_read{0};
-
-	perf_counter_t _comms_errors{perf_alloc(PC_COUNT, MODULE_NAME": com_err")};
-	perf_counter_t _sample_perf{perf_alloc(PC_ELAPSED, MODULE_NAME": read")};
-
-};
+#include "TFMINI.hpp"
 
 TFMINI::TFMINI(const char *port, uint8_t rotation) :
 	ScheduledWorkItem(px4::wq_configurations::hp_default),
@@ -127,7 +62,7 @@ TFMINI::init()
 	case 1: // TFMINI (12m, 100 Hz)
 		_px4_rangefinder.set_min_distance(0.3f);
 		_px4_rangefinder.set_max_distance(12.0f);
-		_px4_rangefinder.set_fov(math::radians(1.15));
+		_px4_rangefinder.set_fov(math::radians(1.15f));
 
 		break;
 
@@ -324,154 +259,4 @@ TFMINI::print_info()
 	perf_print_counter(_comms_errors);
 
 	_px4_rangefinder.print_status();
-}
-
-/**
- * Local functions in support of the shell command.
- */
-namespace tfmini
-{
-
-TFMINI	*g_dev{nullptr};
-
-int start(const char *port, uint8_t rotation);
-int status();
-int stop();
-int usage();
-
-int
-start(const char *port, uint8_t rotation)
-{
-	if (g_dev != nullptr) {
-		PX4_ERR("already started");
-		return PX4_OK;
-	}
-
-	// Instantiate the driver.
-	g_dev = new TFMINI(port, rotation);
-
-	if (g_dev == nullptr) {
-		PX4_ERR("driver start failed");
-		return PX4_ERROR;
-	}
-
-	if (OK != g_dev->init()) {
-		PX4_ERR("driver start failed");
-		delete g_dev;
-		g_dev = nullptr;
-		return PX4_ERROR;
-	}
-
-	return PX4_OK;
-}
-
-int
-status()
-{
-	if (g_dev == nullptr) {
-		PX4_ERR("driver not running");
-		return 1;
-	}
-
-	printf("state @ %p\n", g_dev);
-	g_dev->print_info();
-
-	return 0;
-}
-
-int stop()
-{
-	if (g_dev != nullptr) {
-		PX4_INFO("stopping driver");
-		delete g_dev;
-		g_dev = nullptr;
-		PX4_INFO("driver stopped");
-
-	} else {
-		PX4_ERR("driver not running");
-		return 1;
-	}
-
-	return PX4_OK;
-}
-
-int
-usage()
-{
-	PRINT_MODULE_DESCRIPTION(
-		R"DESCR_STR(
-### Description
-
-Serial bus driver for the Benewake TFmini LiDAR.
-
-Most boards are configured to enable/start the driver on a specified UART using the SENS_TFMINI_CFG parameter.
-
-Setup/usage information: https://docs.px4.io/en/sensor/tfmini.html
-
-### Examples
-
-Attempt to start driver on a specified serial device.
-$ tfmini start -d /dev/ttyS1
-Stop driver
-$ tfmini stop
-)DESCR_STR");
-
-	PRINT_MODULE_USAGE_NAME("tfmini", "driver");
-	PRINT_MODULE_USAGE_SUBCATEGORY("distance_sensor");
-	PRINT_MODULE_USAGE_COMMAND_DESCR("start","Start driver");
-	PRINT_MODULE_USAGE_PARAM_STRING('d', nullptr, nullptr, "Serial device", false);
-	PRINT_MODULE_USAGE_PARAM_INT('R', 25, 1, 25, "Sensor rotation - downward facing by default", true);
-	PRINT_MODULE_USAGE_COMMAND_DESCR("status","Driver status");
-	PRINT_MODULE_USAGE_COMMAND_DESCR("stop","Stop driver");
-	PRINT_MODULE_USAGE_COMMAND_DESCR("test","Test driver (basic functional tests)");
-	PRINT_MODULE_USAGE_COMMAND_DESCR("status","Print driver status");
-	return PX4_OK;
-}
-
-} // namespace
-
-extern "C" __EXPORT int tfmini_main(int argc, char *argv[])
-{
-	int ch = 0;
-	uint8_t rotation = distance_sensor_s::ROTATION_DOWNWARD_FACING;
-	const char *device_path = TFMINI_DEFAULT_PORT;
-	int myoptind = 1;
-	const char *myoptarg = nullptr;
-
-	while ((ch = px4_getopt(argc, argv, "R:d:", &myoptind, &myoptarg)) != EOF) {
-		switch (ch) {
-		case 'R':
-			rotation = (uint8_t)atoi(myoptarg);
-			break;
-
-		case 'd':
-			device_path = myoptarg;
-			break;
-
-		default:
-			PX4_WARN("Unknown option!");
-			return PX4_ERROR;
-		}
-	}
-
-	if (myoptind >= argc) {
-		PX4_ERR("unrecognized command");
-	        return tfmini::usage();
-	}
-
-	if (!strcmp(argv[myoptind], "start")) {
-		if (strcmp(device_path, "") != 0) {
-			return tfmini::start(device_path, rotation);
-
-		} else {
-			PX4_WARN("Please specify device path!");
-			return tfmini::usage();
-		}
-	} else if (!strcmp(argv[myoptind], "stop")) {
-		return tfmini::stop();
-	} else if (!strcmp(argv[myoptind], "status")) {
-		return tfmini::status();
-	}
-
-        return tfmini::usage();
 }
